# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RZNU LAB1",
        "description": "Računarstvo zasnovano na uslugama LAB1",
        "contact": {
            "email": "mislav.zabcic@gmail.com"
        },
        "license": {
            "name": "MIT license",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all posts",
                "description": "Returns array of all posts",
                "operationId": "getAllPosts",
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Creats new post into system",
                "operationId": "create",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreatePost"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error with body parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/posts/{postId}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post by id",
                "description": "Returns a post by his ID",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of a post to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Updates post into system",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "ID of a post to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "Content-type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/x-www-form-urlencoded"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdatePost"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error with body parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post by id",
                "description": "Deletes a post by his ID",
                "operationId": "deletePostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of a post to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/posts/{postId}/image": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post image by id",
                "description": "Returns a image of a post by his ID",
                "operationId": "getPostImageById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of a post image to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succes",
                        "content": {
                            "image/*": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Updates post image in system",
                "operationId": "updatePostImage",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "ID of a post for image to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdatePostImage"
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error with body parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error api response",
                "properties": {
                    "error": {
                        "title": "Error",
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "title": "Post",
                "description": "Post",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "content": {
                        "title": "Content",
                        "type": "string"
                    },
                    "picture": {
                        "title": "Picture"
                    },
                    "user_id": {
                        "title": "User Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "author": {
                        "title": "Author",
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "title": "Created at",
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
            "PostCreate": {
                "title": "Post create",
                "description": "Post create",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "title": "Title",
                        "type": "string"
                    },
                    "content": {
                        "title": "Content",
                        "type": "string"
                    },
                    "picture": {
                        "title": "Picture",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
            "CreatePost": {
                "description": "Object for creating new post",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/PostCreate"
                        }
                    }
                }
            },
            "UpdatePost": {
                "description": "Object for updating post",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/PostUpdate"
                        }
                    }
                }
            },
            "UpdatePostImage": {
                "description": "Object for updating post image",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/PostUpdateImage"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}